requirements.txt
py4j~=0.10.9.7
joblib~=1.3.2
requests~=2.31.0
numpy~=1.26.3
python-dotenv~=1.0.1
pip~=23.3.2
scikit-learn~=1.2.2
pandas~=1.5.3